+++ src/resource/traits.rs
@@ -10,6 +10,10 @@
 
 use super::raw::RawResource;
 use error::WinError;
+use descriptor::{CsuHeap, CbvDesc, SrvBufferDesc, SrvDesc, SrvDimension, UavDimension, UavBufferDesc, UavDesc};
+use device::Device;
+use super::buffer::BufferSlice;
+use pipeline::ia::{IndexBufferView, VertexBufferView};
 
 pub trait Resource {
     /// returns the raw resource reference
@@ -68,6 +72,112 @@ where T: Placed<Heap=H>, H: super::heap::traits::Readback { }
 pub unsafe trait Buffer: Resource {
     /// get the size of the buffer
     fn get_size(&self) -> u64;
+
+    /// Check if this buffer can hold the given `buffer_slice`
+    #[inline]
+    fn is_compatible_with(&self, buffer_slice: BufferSlice) -> bool {
+        (buffer_slice.offset + buffer_slice.length as u64) * buffer_slice.byte_stride as u64 <= self.get_size()
+    }
+
+    /// create a cbv for this buffer on `csu_heap` at `index`
+    #[inline]
+    fn create_cbv<DH: CsuHeap>(
+        &mut self, device: &mut Device, csu_heap: &mut DH, index: u32
+    ) {
+        csu_heap.create_cbv(device, &CbvDesc{
+            buffer_location: self.as_raw_mut().get_gpu_vaddress(),
+            size: self.get_size() as u32, // TODO: is this safe?
+        }, index);
+    }
+
+    // TODO: investigate rtv on buffer, https://msdn.microsoft.com/en-us/library/windows/desktop/dn770342(v=vs.85).aspx
+
+    /// Create a srv for this buffer on `csu_heap` at `index` with the given `slice`
+    /// TODO: verify, raw buffer support
+    #[inline]
+    fn create_srv<DH: CsuHeap>(
+        &mut self, device: &mut Device, csu_heap: &mut DH,
+        index: u32, slice: BufferSlice
+    ) {
+        debug_assert!(self.is_compatible_with(slice));
+        csu_heap.create_srv(device, Some(self.as_raw()), Some(&SrvDesc{
+            format: ::format::DXGI_FORMAT_UNKNOWN, // TODO: double check format for buffer
+            dimension: SrvDimension::Buffer(SrvBufferDesc{
+                offset: slice.offset, num_elements: slice.length,
+                byte_stride: slice.byte_stride, raw: 0 // TODO: support raw buffers?
+            }),
+            component_mapping: Default::default(), // TODO: would different component mapping make sense here?
+        }), index);
+    }
+
+    /// Create a uav for this buffer on `csu_heap` at `index` with the given `slice`
+    /// TODO: verify
+    #[inline]
+    fn create_uav<DH: CsuHeap, B: AllowUnorderedAccess + Buffer>(
+        buf: &mut B, device: &mut Device, csu_heap: &mut DH,
+        index: u32, slice: BufferSlice
+    ) {
+        debug_assert!(buf.is_compatible_with(slice));
+        csu_heap.create_uav(device, Some(buf.as_raw()), None, Some(&UavDesc{
+            format: ::format::DXGI_FORMAT_UNKNOWN, // TODO: double check format for buffer
+            dimension: UavDimension::Buffer(UavBufferDesc{
+                offset: slice.offset, num_elements: slice.length,
+                byte_stride: slice.byte_stride, 
+                counter_offset: 0,
+                raw: 0 // TODO: support raw buffers?
+            })
+        }), index);
+    }
+
+    /// Create a uav for this buffer on `csu_heap` at `index` with the given `slice` and `counter`
+    /// TODO: verify
+    #[inline]
+    fn create_uav_with_counter<DH: CsuHeap, CounterBuf: Buffer, B: AllowUnorderedAccess + Buffer>(
+        buf: &mut B, device: &mut Device, csu_heap: &mut DH,
+        index: u32, slice: BufferSlice, counter: &mut CounterBuf,
+        counter_offset: u64
+    ) {
+        debug_assert!(buf.is_compatible_with(slice));
+        debug_assert!(counter_offset%4 == 0);
+        debug_assert!(counter_offset <= counter.get_size());
+        csu_heap.create_uav(device, Some(buf.as_raw()), Some(counter.as_raw()), Some(&UavDesc{
+            format: ::format::DXGI_FORMAT_UNKNOWN, // TODO: double check format for buffer
+            dimension: UavDimension::Buffer(UavBufferDesc{
+                offset: slice.offset, num_elements: slice.length,
+                byte_stride: slice.byte_stride, 
+                counter_offset, raw: 0 // TODO: support raw buffers?
+            })
+        }), index);
+    }
+
+    /// Create a vbv for this buffer
+    #[inline]
+    fn create_vbv(&mut self, size: u32, stride: u32) -> VertexBufferView {
+        debug_assert!(size as u64 <= self.get_size());
+        VertexBufferView{
+            location: self.as_raw_mut().get_gpu_vaddress(), size, stride
+        }
+    }
+
+    /// Create an ibv for this buffer
+    #[inline]
+    fn create_ibv_u32(&mut self, size: u32) -> IndexBufferView {
+        debug_assert!(size as u64 <= self.get_size());
+        IndexBufferView{
+            location: self.as_raw_mut().get_gpu_vaddress(), size,
+            format: ::format::DXGI_FORMAT_R32_UINT,
+        }
+    }
+
+    /// Create an ibv for this buffer
+    #[inline]
+    fn create_ibv_u16(&mut self, size: u32) -> IndexBufferView {
+        debug_assert!(size as u64 <= self.get_size());
+        IndexBufferView{
+            location: self.as_raw_mut().get_gpu_vaddress(), size,
+            format: ::format::DXGI_FORMAT_R16_UINT,
+        }
+    }
 }
 
 /// a cpu-writable buffer